# docs/openapi.yaml
basePath: /api
...
host: localhost:8090
info:
  contact: {}
  description: Self-hosted CTF platform API. Most write endpoints require authentication
    via JWT cookie (auth_token).
  title: hCTF API
  version: 1.0.0
paths:
...
  /auth/forgot-password:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Generates a reset token and sends email via configured SMTP.
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: If that email exists, a reset link has been sent.
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Request a password reset token
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              token:
                type: string
              user:
                type: object
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Authenticate a user and set session cookie
      tags:
      - Auth
  /auth/logout:
    post:
      produces:
      - text/plain
      responses:
        "200":
          description: Logged out
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Clear authentication cookie and end session
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: User email
        in: formData
        name: email
        required: true
        type: string
      - description: User password
        in: formData
        name: password
        required: true
        type: string
      - description: Display name
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              token:
                type: string
              user:
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Register a new user account
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: Reset token
        in: formData
        name: token
        required: true
        type: string
      - description: New password
        in: formData
        name: password
        required: true
        type: string
      - description: Confirm new password
        in: formData
        name: confirm_password
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Password reset successful. You can now login.
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Reset password using a valid reset token
      tags:
      - Auth
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: auth_token
    type: apiKey
swagger: "2.0"